requirements.txt
action-msgs==1.2.2
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.9.0
ament-cmake-test==1.3.12
ament-copyright==0.12.14
ament-cppcheck==0.12.14
ament-cpplint==0.12.14
ament-flake8==0.12.14
ament-index-python==1.4.0
ament-lint==0.12.14
ament-lint-cmake==0.12.14
ament-package==0.14.1
ament-pep257==0.12.14
ament-uncrustify==0.12.14
ament-xmllint==0.12.14
angles==1.15.0
annotated-types==0.7.0
anyio==4.11.0
builtin-interfaces==1.2.2
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
composition-interfaces==1.2.2
control-msgs==4.8.0
controller-manager==2.51.0
controller-manager-msgs==2.51.0
cv-bridge==3.2.1
demo-nodes-py==0.20.5
diagnostic-msgs==4.9.0
domain-coordinator==0.10.0
example-interfaces==0.9.3
examples-rclpy-executors==0.15.4
examples-rclpy-minimal-action-client==0.15.4
examples-rclpy-minimal-action-server==0.15.4
examples-rclpy-minimal-client==0.15.4
examples-rclpy-minimal-publisher==0.15.4
examples-rclpy-minimal-service==0.15.4
examples-rclpy-minimal-subscriber==0.15.4
exceptiongroup==1.3.0
fastapi==0.119.0
geometry-msgs==4.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httplib2==0.31.0
idna==3.11
image-geometry==3.2.1
interactive-markers==2.3.2
laser-geometry==2.4.0
launch==1.0.10
launch-ros==0.19.10
launch-testing==1.0.10
launch-testing-ros==0.19.10
launch-xml==1.0.10
launch-yaml==1.0.10
lifecycle-msgs==1.2.2
logging-demo==0.20.5
map-msgs==2.1.0
message-filters==4.3.8
mocap4r2-control-msgs==0.0.7
nav-msgs==4.9.0
numpy==2.2.6
osrf-pycommon==2.1.6
packaging==25.0
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
pyparsing==3.2.5
python-dotenv==1.1.1
python-qt-binding==1.1.2
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.2
rclpy==3.3.17
rcutils==5.1.6
requests==2.32.5
resource-retriever==3.1.3
rmw-dds-common==1.6.0
ros2action==0.18.13
ros2bag==0.15.15
ros2cli==0.18.13
ros2component==0.18.13
ros2doctor==0.18.13
ros2interface==0.18.13
ros2launch==0.19.10
ros2lifecycle==0.18.13
ros2multicast==0.18.13
ros2node==0.18.13
ros2param==0.18.13
ros2pkg==0.18.13
ros2run==0.18.13
ros2service==0.18.13
ros2topic==0.18.13
rosbag2-interfaces==0.15.15
rosbag2-py==0.15.15
rosgraph-msgs==1.2.2
rosidl-adapter==3.1.7
rosidl-cli==3.1.7
rosidl-cmake==3.1.7
rosidl-generator-c==3.1.7
rosidl-generator-cpp==3.1.7
rosidl-generator-py==0.14.5
rosidl-parser==3.1.7
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.7
rosidl-typesupport-introspection-cpp==3.1.7
rpyutils==0.2.1
rqt==1.1.7
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-controller-manager==2.51.0
rqt-dotgraph==0.0.4
rqt-gauges==0.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-joint-trajectory-controller==2.49.1
rqt-moveit==1.0.1
rqt-msg==1.2.0
rqt-plot==1.1.5
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==1.0.6
rqt-robot-steering==1.0.3
rqt-runtime-monitor==1.0.0
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-tf-tree==1.0.5
rqt-topic==1.5.0
rsa==4.9.1
sensor-msgs==4.9.0
sensor-msgs-py==4.9.0
shape-msgs==4.9.0
sniffio==1.3.1
sros2==0.10.7
starlette==0.48.0
statistics-msgs==1.2.2
std-msgs==4.9.0
std-srvs==4.9.0
stereo-msgs==4.9.0
teleop-twist-keyboard==2.4.0
tf2-geometry-msgs==0.25.16
tf2-kdl==0.25.16
tf2-msgs==0.25.16
tf2-py==0.25.16
tf2-ros-py==0.25.16
tf2-sensor-msgs==0.25.16
tf2-tools==0.25.16
topic-monitor==0.20.5
tqdm==4.67.1
trajectory-msgs==4.9.0
turtlesim==1.4.2
typing-inspection==0.4.2
typing_extensions==4.15.0
unique-identifier-msgs==2.2.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
visualization-msgs==4.9.0
websockets==15.0.1
